Moved AppState / Services to its own library
