+++ projects/app-state/ng-package.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../dist/app-state",
+  "lib": {
+    "entryFile": "src/projects.ts"
+  }
+}
