+++ projects/app-state/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "app-state",
+  "version": "0.0.1",
+  "peerDependencies": {
+    "@angular/common": "^11.2.9",
+    "@angular/core": "^11.2.9"
+  },
+  "dependencies": {
+    "tslib": "^2.0.0"
+  }
+}
