+++ projects/app-state/src/lib/services/network-interfaces.service.ts
@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
 import {ENDPOINTS, NetworkInfo} from "@memebox/contracts";
-import {EXPRESS_BASE} from "../../state/app.service";
+import {EXPRESS_BASE} from "../state/app.service";
 import {map, shareReplay} from "rxjs/operators";
 import {Observable} from "rxjs";
