+++ projects/app-state/src/lib/state/index.ts
@@ -1,0 +1,2 @@
+export * from './app.queries'
+export * from './app.store';
