+++ projects/app-state/src/projects.ts
@@ -1,0 +1,7 @@
+/*
+ * Public API Surface of app-state
+ */
+
+export * from './lib/services.module';
+export * from './lib/services';
+export * from './lib/state';
